+++ photos/photos.resolvers.js
@@ -13,5 +13,27 @@ export default {
           },
         },
       }),
+    likes: ({ id }) => client.like.count({ where: { photoId: id } }),
+  },
+  Hashtag: {
+    photos: ({ id }, { page }) =>
+      client.hashtag
+        .findUnique({
+          where: { id },
+        })
+        .photos({
+          take: 5,
+          skip: (page - 1) * 5,
+        }),
+    totalPhotos: ({ id }) =>
+      client.photo.count({
+        where: {
+          hashtags: {
+            some: {
+              id,
+            },
+          },
+        },
+      }),
   },
 };