+++ photos/photos.utils.js
@@ -1,0 +1,9 @@
+export const processHashtags = (caption) => {
+  //parse caption
+  const hashtags = caption.match(/#[\w]+/g) || [];
+  //get or create Hashtags
+  return hashtags.map((hashtag) => ({
+    where: { hashtag },
+    create: { hashtag },
+  }));
+};