+++ users/users.typeDefs.js
@@ -11,6 +11,7 @@ export default gql`
     updatedAt: String!
     bio: String
     avatar: String
+    photos: [Photo]
     following: [User]
     followers: [User]
     totalFollowing: Int!
@@ -19,6 +20,3 @@ export default gql`
     isFollowing: Boolean!
   }
 `;
-
-//
-//